refactor: split deploy prepare runtime contract loading and finalization
Move deployment runtime contract loading and previous-phase AAC fallback out of operations.ts into runtime-contract.ts so deploy prepare orchestration no longer owns delivery locator and architecture artifact lookup details.

Consolidate repeated deploy prepare finalization across compose-existing, dockerfile-existing, and generated Dockerfile providers so spec writing, failure cleanup, prepare logging, and result shaping all pass through one helper.

Verified with npm run build, npm run verify:deploy-structured-blockers, and npm run smoke:deploy.
